requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
et_xmlfile==2.0.0
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pillow==12.2.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
